Synchronize the archive requester queue with responses
This change re-establishes a 1 to 1 relationship between requester
and response in the archive.

Before this we didn't remove the entries from the responders queue
if the same client had multiple ongoing requests. Multiple identical
entries in the queue would only be cleaned up after all requests
from a client were fulfilled.